Zomboid server is dead, fucking R.I.P. Reasoning: I've followed Zomboid since it's alpha state in 2011. I originally purchased the game on Desura (hey remember that!) and since then, it seems like it hasn't progressed.
The game still runs like shit (both for servers and clients), maps/mods
that are popular and in use currently are unmaintained, and it's just
annoying getting anything to work properly.

The game lacks any form of well-structured documentation, and help from
the developers is basically non-existent.

The online community is also pretty dead.

The playerbase has surges during new updates, but quickly dies down
afterwards.

I tried really hard to keep things running smoothly, but ultimately
things just didn't work out.

I hope one day the issues get resolved. But, for now, I'm giving up on
it.

If you disagree with this decision, you can contact me on the Discord
and talk me out of it, or on Twitter.

-Keith