Did a complete redesign of the wait:  The goal was to make the site much more appealing and simplistic in  design.
The new layout has issues on mobile which I hope to sort out soon - I'm
still working on the new design and by no means is it completely
finished.

There's a lot of ugly copy/pasting in the stylesheet which I hope I can
condense.

Once I think the site is finished in both its functionality and code
quality, I will tag it as V3 of the site and I will go back and tag the
old version as V2