Removed the zomboid server for good.
This marks the last release of the current website design, as I'm
growing sick of it and want to create a more personalized page.

Life has been hitting me pretty hard lately, and I find it difficult to
aspire myself to work on hobby projects. But slowly I've been convincing
myself that, for my sanity, it's best to keep myself busy with
something.

It's hard to say if any of the game servers will stay up as I really
don't play much games anymore and not in the interest of hosting game
servers anymore.